+++ css/mby-lab-article.css
@@ -1,0 +1,27 @@
+.page-path div:nth-child(2n){
+  margin: 0 10px 0 10px;
+}
+
+.pdf-container {
+  width: 100%;
+  height: 90%;
+  margin-top: 20px;
+}
+
+.article-pdf {
+  width: 100%;
+  height: 100%;
+}
+
+.notice-container {
+  padding-top: 8%;
+  width: 100%;
+  height: 100%;
+  text-align: center;
+  margin: 100px auto;
+  font-size: 30px;
+}
+
+body {
+  max-height: 450px;
+}
