+++ css/mby-lab-weeklyReports.css
@@ -33,17 +33,52 @@
 .search-button {
   vertical-align: baseline;
   text-align: center;
-  width: 120px;
+  width: 140px;
   height: 30px;
   margin-left: 20px;
   border-style: solid;
   border-width: 1px;
+  font-size: 15px;
+  padding-top: 3px;
 }
 
 .search-button:hover {
   background-color: #202649;
   color: #49BAA9;
   transition: all 0.3s;
+  cursor: pointer;
+}
+
+.modal {
+  display: none;
+  position: fixed;
+  z-index: 1;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: #000;
+  background-color: rgba(0,0,0,0.4);
+}
+
+.modal-content {
+  background-color: #fefefe;
+  margin: 15% auto;
+  padding: 20px;
+  border: 1px solid #888;
+  width: 50%;
+}
+
+.close {
+  color: #aaa;
+  float: right;
+  font-size: 28px;
+  font-weight: bold;
+}
+
+.close:hover, .close:focus {
+  color: black;
+  text-decoration: none;
   cursor: pointer;
 }
 
@@ -93,9 +128,26 @@
 }
 
 .article-image {
+  position: relative;
   width: 100%;
   height: 250px;
   background-color: #00008f;
+}
+
+.article-image h5:hover {
+  color: #49BAA9;
+  transition: all 0.3s;
+  cursor: pointer;
+}
+
+.article-image h5 {
+  position: absolute;
+  top: 0;
+  left: 0;
+  padding-left: 50px;
+  color: #fff;
+  font-size: 30px;
+  font-weight: 350;
 }
 
 .weekly-article-text {
@@ -107,4 +159,19 @@
 .article-box h6 {
   margin-left: 20px;
   font-size: 12px;
+}
+
+.page-path div:nth-child(2n){
+  margin: 0 10px 0 10px;
+}
+
+.pdf-container {
+  width: 100%;
+  height: 90%;
+  margin-top: 20px;
+}
+
+.article-pdf {
+  width: 100%;
+  height: 100%;
 }
