+++ reportsArticle.html
@@ -1,0 +1,134 @@
+<!DOCTYPE html>
+<html lang="en">
+
+  <head>
+
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta name="description" content="">
+    <meta name="author" content="">
+
+    <title>MBY LAB | Weekly Reports</title>
+
+    <!-- Bootstrap core CSS -->
+    <link href="vendor/bootstrap/css/bootstrap.min.css" rel="stylesheet">
+
+    <!-- Carousel CSS and JavaScript-->
+    <link href="carousel.css" rel="stylesheet">
+    <script src="carousel.min.js" type="text/javaScript"></script>
+
+    <!-- Custom fonts for this template -->
+    <link href="vendor/fontawesome-free/css/all.min.css" rel="stylesheet">
+    <link rel="stylesheet" href="vendor/simple-line-icons/css/simple-line-icons.css">
+    <link href="https://fonts.googleapis.com/css?family=Lato" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css?family=Catamaran:100,200,300,400,500,600,700,800,900" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css?family=Muli" rel="stylesheet">
+
+    <!-- Custom styles for this template -->
+    <link href="css/mby-lab.css" rel="stylesheet">
+    <link href="css/mby-lab-weeklyReports.css" rel="stylesheet">
+
+    <!-- Add Favicon -->
+    <link rel="shortcut icon" href="img/mby_favicon.png">
+
+
+  </head>
+
+  <body id="page-top">
+    <div class="fixed-top">
+    <!-- TradingView Widget BEGIN -->
+    <div class="tradingview-widget-container">
+      <div class="tradingview-widget-container__widget"></div>
+      <script type="text/javascript" class="widget-ticker" src="https://s3.tradingview.com/external-embedding/embed-widget-ticker-tape.js" async>
+      {
+        "symbols": [
+          {
+            "title": "S&P 500",
+            "proName": "OANDA:SPX500USD"
+          },
+          {
+            "title": "나스닥 100",
+            "proName": "OANDA:NAS100USD"
+          },
+          {
+            "title": "EUR/USD",
+            "proName": "FX_IDC:EURUSD"
+          },
+          {
+            "title": "BTC/USD",
+            "proName": "BITSTAMP:BTCUSD"
+          },
+          {
+            "title": "ETH/USD",
+            "proName": "BITSTAMP:ETHUSD"
+          }
+        ],
+        "colorTheme": "light",
+        "isTransparent": false,
+        "displayMode": "adaptive",
+        "locale": "kr"
+      }
+      </script>
+    </div>
+    <!-- TradingView Widget END -->
+      <!-- Navigation -->
+      <nav class="navbar navbar-expand-lg" id="mainNav">
+        <div class="nav-container container">
+          <a class="navbar-brand js-scroll-trigger" href="#page-top"></a>
+          <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarResponsive" aria-controls="navbarResponsive" aria-expanded="false" aria-label="Toggle navigation">
+            Menu
+            <i class="fas fa-bars"></i>
+          </button>
+          <a class="mby_logo" href="index.html">
+            <img src="img/mby_lab_logo.png">
+          </a>
+          <div class="collapse navbar-collapse" id="navbarResponsive">
+            <ul class="navbar-nav ml-auto">
+              <li class="nav-item">
+                <a class="nav-link js-scroll-trigger" href="index.html">Home</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link js-scroll-trigger" href="weeklyReports.html">Weekly Reports</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link js-scroll-trigger">Blog</a>
+              </li>
+              <li class="nav-item">
+                <a class="nav-link js-scroll-trigger">About</a>
+              </li>
+            </ul>
+          </div>
+        </div>
+      </nav>
+    </div>
+    <div class="page-path search-bar container">
+      <div><a href="weeklyReports.html">Weekly Reports</a></div>
+      <div>></div>
+      <div>Incoming & Saving</div>
+    </div>
+    <div class="container weeklyReports-main-wrapper">
+      <div class="weeklyReports-main">
+        <img src="img/weeklyReports_headImg.png"></img>
+        <div class="main-contents">
+          <h1>Weekly Reports</h1>
+          <h5>Explore the forum from MBY LAB</h5>
+        </div>
+      </div>
+    </div>
+    <div class="container pdf-container">
+      <iframe class="article-pdf" src="img/finports_w_report.pdf"></iframe>
+    </div>
+    <footer>
+      <div class="container">
+        <p>Copyright &copy; 2019 <a href="https://finports.com">FinPorts Inc.</a> All Rights Reserved.</p>
+    </footer>
+    <!-- Bootstrap core JavaScript -->
+    <script src="vendor/jquery/jquery.min.js"></script>
+    <script src="vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
+
+    <!-- Plugin JavaScript -->
+    <script src="vendor/jquery-easing/jquery.easing.min.js"></script>
+
+    <script src="js/mby_lab.js"></script>
+    <script src="js/mby_lab_search.js"></script>
+</html>
