--- slick-1.8.1/.gitignore
@@ -1,2 +1,0 @@
-.sass-cache
-node_modules
