--- slick-1.8.1/Makefile
@@ -1,8 +1,0 @@
-
-build: components slick.js slick.css
-  @component build --dev
-
-components: component.json
-  @component install --dev
-
-.PHONY: clean
