--- slick-1.8.1/bower.json
@@ -1,32 +1,0 @@
-{
-    "name": "slick-carousel",
-    "main": [
-        "slick/slick.js",
-        "slick/slick.css",
-        "slick/slick.less",
-        "slick/slick.scss"
-    ],
-    "homepage": "https://github.com/kenwheeler/slick",
-    "authors": [
-        "Ken Wheeler <ken_wheeler@me.com>"
-    ],
-    "description": "the last carousel you'll ever need",
-    "keywords": [
-        "responsive",
-        "carousel",
-        "jquery"
-    ],
-    "license": "MIT",
-    "ignore": [
-        "**/.*",
-        "node_modules",
-        "bower_components",
-        "test",
-        "tests",
-        "package.json",
-        "index.html"
-    ],
-    "dependencies": {
-        "jquery": ">=1.7"
-    }
-}
