--- slick-1.8.1/component.json
@@ -1,18 +1,0 @@
-{
-    "name": "slick",
-    "repo": "kenwheeler/slick",
-    "description": "the last carousel you'll ever need",
-    "version": "1.8.0",
-    "keywords": ["ui", "jquery", "carousel", "responsive", "slider"],
-    "dependencies": {
-        "component/jquery": "*"
-    },
-    "development": {},
-    "scripts": {
-        "slick/slick.js": "slick.js"
-    },
-    "styles": {
-        "slick/slick.css": "slick.css",
-        "slick/slick-theme.css": "slick-theme.css"
-    }
-}
