--- slick-1.8.1/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "slick-carousel",
-  "version": "1.8.1",
-  "description": "the last carousel you'll ever need",
-  "main": "slick/slick.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/kenwheeler/slick.git"
-  },
-  "keywords": [
-    "carousel",
-    "slick",
-    "responsive",
-    "js",
-    "slider"
-  ],
-  "author": "Ken Wheeler <ken_wheeler@me.com>",
-  "contributors": [
-    {
-      "name": "ken wheeler",
-      "email": "ken_wheeler@me.com",
-      "url": "http://kenwheeler.github.io/"
-    },
-    {
-      "name": "simon goellner",
-      "email": "simey.me@gmail.com",
-      "url": "http://simey.me"
-    },
-    {
-      "name": "ahmad el-alfy",
-      "email": "ahmadalfy@gmail.com",
-      "url": "http://www.alfy.me"
-    },
-    {
-      "name": "leggomuhgreggo",
-      "email": "gwestneat@gmail.com"
-    },
-    {
-      "name": "ashley mcknight",
-      "email": "ash@sharpteef.net"
-    }
-  ],
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/kenwheeler/slick/issues"
-  },
-  "peerDependencies": {
-    "jquery": ">=1.8.0"
-  }
-}
