--- slick-1.8.1/slick.jquery.json
@@ -1,34 +1,0 @@
-{
-    "name": "slick",
-    "title": "slick",
-    "description": "A jQuery responsive touch carousel",
-    "keywords": [
-        "responsive",
-        "slider",
-        "carousel",
-        "touch",
-        "mobile"
-    ],
-    "version": "1.8.0",
-    "author": {
-        "name": "Ken Wheeler",
-        "url": "http://kenwheeler.github.io"
-    },
-    "maintainers": [{
-        "name": "Ken Wheeler",
-        "email": "dubmediagroup@gmail.com",
-        "url": "http://www.dubmediagroup.com"
-    }],
-    "licenses": [{
-        "type": "MIT",
-        "url": "https://github.com/kenwheeler/slick/blob/master/LICENSE"
-    }],
-    "demo": "http://kenwheeler.github.io/slick/",
-    "bugs": "https://github.com/kenwheeler/slick/issues",
-    "homepage": "https://github.com/kenwheeler/slick/",
-    "docs": "https://github.com/kenwheeler/slick/",
-    "download": "https://github.com/kenwheeler/slick/archive/master.zip",
-    "dependencies": {
-        "jquery": ">=1.7"
-    }
-}
